Refactor convert_kursliste_to_sqlite and integrate into CLI
This change refactors the logic from
`scripts/convert_kursliste_to_sqlite.py` into a new module
`src/opensteuerauszug/kursliste/converter.py` and integrates it into the
main CLI.

Key changes:

- Created `src/opensteuerauszug/kursliste/converter.py` containing the
  logic to convert Kursliste XML files to SQLite.
- Updated `src/opensteuerauszug/kursliste/__main__.py`:
    - Added a `convert` command to manually trigger conversion.
    - Updated `download` command to include a `--convert/--no-convert`
      flag (default True), automatically converting downloaded files.
- Updated `scripts/convert_kursliste_to_sqlite.py` to import and use the
  new module, preserving backward compatibility for the script.
- Added tests for the new CLI commands in `tests/kursliste/`.
- Updated documentation to reflect the new automated download and
  conversion workflow.

This improves the workflow by allowing users to download and index
Kursliste data in one step and provides a standard CLI command for
conversion.
